+++ WSTransport/src/main/java/data_types/StopPoint.java
@@ -1,15 +1,15 @@
 package data_types;
 
 public class StopPoint {
-	private final int mId;
+	private final long mId;
 	private final String mFriendlyName;
 	
-	public StopPoint(int id, String friendlyName) {
+	public StopPoint(long id, String friendlyName) {
 		mId = id;
 		mFriendlyName = friendlyName;
 	}
 	
-	public int getId() {
+	public long getId() {
 		return mId;
 	}
